Close dialog on publish, move thing around
